+++ src/index.js
@@ -2,7 +2,7 @@ import './js/init-home';
 import './js/search-bar';
 import './js/fetch-modal';
 import './js/modal-film';
-import './js/fireBase/authorization';
+import './js/authorization';
 import './js/fireBase/firebase-auth';
 //------Удалити-----//
 import './js/fireBase/firestore';
