+++ src/js/fireBase/firestore.js
@@ -1,53 +1,53 @@
 import { initDataBase } from './utils';
-import { addDoc, getDocs, collection } from 'firebase/firestore';
-import { async } from '@firebase/util';
+import { addDoc, getDocs, collection, doc } from 'firebase/firestore';
+import { getAuth, onAuthStateChanged } from 'firebase/auth';
+import { refs } from '../refs';
 
+refs.base.addEventListener('click', addDataWatched);
+
+//
 const db = initDataBase();
+const auth = getAuth();
 
-//------------Function for ADD to firestore---------------------//
+// //------------Function for ADD to firestore---------------------//
 
 async function addDataWatched(object) {
   try {
-    const docRef = await addDoc(collection(db, 'watched'), object);
-    docRef.id;
+    const docRef = await addDoc(
+      collection(db, 'filmoteka', auth.currentUser.email, 'watched'),
+      object
+    );
+    console.log(docRef.id);
+    console.log('FUNC DATA added to WATCHED');
+  } catch (e) {
+    console.error('Error adding document: ', e);
+  }
+}
+
+async function addDataQueue(object) {
+  try {
+    const docRef = await addDoc(
+      collection(db, 'filmoteka', auth.currentUser.email, 'queue'),
+      object
+    );
     console.dir(docRef);
   } catch (e) {
     console.error('Error adding document: ', e);
   }
 }
-async function addDataQueue(object) {
-  try {
-    const docRef = await addDoc(collection(db, 'queue'), object);
-    console.log('Document written with ID: ', docRef.id);
-  } catch (e) {
-    console.error('Error adding document: ', e);
-  }
-}
 
-addDataWatched({
-  first: 1,
-  middle: 'Sasha',
-  last: 'gddhd',
-  born: 1984,
-});
-
-addDataQueue({
-  first: 3,
-  middle: 'Queue',
-  last: 'gddhd',
-  born: 1984,
-});
-
-//------------Function for GET from firestore---------------------//
+// //------------Function for GET from firestore---------------------//
 async function getDataWatched() {
   const response = [];
-  const querySnapshot = await getDocs(collection(db, 'watched'));
+  const querySnapshot = await getDocs(collection(db, 'filmoteca'));
+  console.log(querySnapshot);
   querySnapshot.forEach(doc => {
-    console.log(`doc.data()`);
     response.push(doc.data());
   });
+  console.log('response', response);
   return response;
 }
 
-const data = getDataWatched();
-console.log('fuck', data);
+// const data = getDataWatched();
+
+// console.log(data);
