+++ src/js/refs.js
@@ -12,4 +12,31 @@ export const refs = {
   students: document.querySelector('.students'),
   footerBtn: document.querySelector('.footer-btn'),
   backdropFooter: document.querySelector('.backdrop-footer'),
+  openModalAuthBtn: document.querySelector('[data-open-modal]'),
+  signOutBtn: document.querySelector('[data-sign-out]'),
+  sigInGoogleBtn: document.querySelector('[data-sigin-google]'),
+  modal: document.querySelector('[data-modal-auth]'),
+  body: document.querySelector('body'),
+  closeModalAuthBtn: document.querySelector('[data-modal-auth-close]'),
+  userInfo: document.querySelector('.heder_auth_info'),
+  authSwitchLinks: document.querySelectorAll('.switch'),
+  authModals: document.querySelectorAll('.auth .modal-auth'),
+  authWrapper: document.querySelector('.auth'),
+  registerForm: document.querySelector('.register'),
+  loginForm: document.querySelector('.login'),
+
+  openModalAuthBtn: document.querySelector('[data-open-modal]'),
+  //--------- потім удалити---------------////
+  base: document.querySelector('[data-open-base]'),
+  signOutBtn: document.querySelector('[data-sign-out]'),
+  sigInGoogleBtn: document.querySelector('[data-sigin-google]'),
+  modal: document.querySelector('[data-modal-auth]'),
+  body: document.querySelector('body'),
+  closeModalAuthBtn: document.querySelector('[data-modal-auth-close]'),
+  userInfo: document.querySelector('.heder_auth_info'),
+  authSwitchLinks: document.querySelectorAll('.switch'),
+  authModals: document.querySelectorAll('.auth .modal-auth'),
+  authWrapper: document.querySelector('.auth'),
+  registerForm: document.querySelector('.register'),
+  loginForm: document.querySelector('.login'),
 };
