+++ src/partials/authorization.html
@@ -22,6 +22,9 @@
           placeholder="Password"
         />
         <button type="submit" class="modal-btn" id="login-btn">Login</button>
+        <button type="button" class="modal-btn btn-auth" data-sigin-google>
+          Sign In with Google
+        </button>
         <p class="error"></p>
       </form>
       <div>No account? <button class="switch">Register instead</button></div>
