+++ src/partials/header.html
@@ -29,9 +29,6 @@
         <button type="button" class="btn-auth is-hidden" data-sign-out>
           Sign Out
         </button>
-        <button type="button" class="btn-auth" data-sigin-google>
-          Sign In with Google
-        </button>
       </div>
     </nav>
